refactor(trusty-common): split memory_tui into submodules under 500-SLOC cap (#607)
Split the 2668-SLOC monolith crates/trusty-common/src/monitor/memory_tui.rs
into a focused module directory, bringing every production file under the
500-SLOC gate enforced by check_line_cap.sh:

  activity.rs    —  92 SLOC  (spinners, PalaceActivity, timing helpers)
  state.rs       — 259 SLOC  (MemoryFocus, DreamBackoff, DrawerListState, MemoryTuiState)
  view.rs        — 422 SLOC  (palace_lines, stats_lines, title_line, drawer helpers, navigation)
  render.rs      — 287 SLOC  (run_with_url, render, render_activity_and_stats, render_detail_pane)
  event_loop.rs  — 362 SLOC  (poll_daemon, run_recall, apply_memory_event, fetch_drawer_*, run_loop)
  mod.rs         —  29 SLOC  (PalaceSortKey alias, pub re-exports, run() entry point)
  tests.rs       — 1286 SLOC (full test suite; classified as test file → 1500-SLOC cap)

Public API and all 219 tests are unchanged. The allowlist entry for the old
memory_tui.rs (budget 2668) is removed by `check_line_cap.sh --update`.
