+++ .line-cap-allowlist.tsv
@@ -26,7 +26,6 @@ crates/trusty-common/src/memory_core/store/kg.rs	815
 crates/trusty-common/src/memory_core/store/payload_store.rs	547
 crates/trusty-common/src/monitor/dashboard.rs	745
 crates/trusty-common/src/monitor/memory_client.rs	754
-crates/trusty-common/src/monitor/memory_tui.rs	2668
 crates/trusty-common/src/monitor/search_tui.rs	1596
 crates/trusty-common/src/tickets/api/backends/github.rs	721
 crates/trusty-common/src/tickets/api/backends/jira.rs	758
